pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "nanolfa-design"
version = "0.1.0"
description = "Iterative AlphaFold-guided nanobody design for lateral flow immunoassays targeting urinary hormones"
readme = "README.md"
license = {text = "Apache-2.0"}
requires-python = ">=3.10"
authors = [
    {name = "DoctorDean"},
]
keywords = [
    "nanobody",
    "alphafold",
    "lateral-flow-assay",
    "antibody-design",
    "structure-prediction",
    "protein-engineering",
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Scientific/Engineering :: Bio-Informatics",
    "Topic :: Scientific/Engineering :: Chemistry",
]

dependencies = [
    "numpy>=1.26",
    "scipy>=1.13",
    "pandas>=2.2",
    "biopython>=1.83",
    "biotite>=0.41",
    "rdkit>=2024.03",
    "torch>=2.3",
    "omegaconf>=2.3",
    "hydra-core>=1.3",
    "click>=8.1",
    "rich>=13.7",
    "pyyaml>=6.0",
    "freesasa>=2.2",
    "fair-esm>=2.0",
    "matplotlib>=3.9",
    "wandb>=0.17",
    "mdtraj>=1.9",
    "scikit-learn>=1.5",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.2",
    "pytest-cov>=5.0",
    "mypy>=1.10",
    "ruff>=0.4",
]
notebooks = [
    "jupyterlab>=4.2",
    "ipywidgets>=8.1",
    "nglview>=3.1",
    "plotly>=5.22",
    "seaborn>=0.13",
]
hpc = [
    "dask>=2024.5",
    "ray>=2.10",
    "submitit>=1.5",
]

[project.scripts]
nanolfa = "nanolfa.core.pipeline:main"

[project.urls]
Repository = "https://github.com/DoctorDean/nanobody-lfa-design"
Documentation = "https://github.com/DoctorDean/nanobody-lfa-design/tree/main/docs"

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
target-version = "py310"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "B", "A", "SIM"]

[tool.mypy]
python_version = "3.10"
warn_return_any = false
warn_unused_configs = true
disallow_untyped_defs = false
ignore_missing_imports = true
# Scientific libs (Biopython, FreeSASA, PyRosetta, RDKit) have incomplete
# type stubs. Strict mode here would fail for reasons unrelated to our code.
# Tighten these as we add typed wrappers around third-party calls.

[[tool.mypy.overrides]]
module = "nanolfa.utils.chemistry"
disable_error_code = ["assignment", "index", "arg-type", "attr-defined"]

[[tool.mypy.overrides]]
module = "nanolfa.models.md_validation"
disable_error_code = ["operator", "var-annotated"]

[[tool.mypy.overrides]]
module = "nanolfa.utils.plotting"
disable_error_code = ["attr-defined"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v"
